refactor: Modularise campaign components
