+++ components/CampaignSegment/FeaturedCampaignsContainer/CampaignPanel.tsx
